Move Edge to the correct namespace
